+++ src/data/navigaton-data.js
@@ -1,18 +1,21 @@
 
+/**
+ * An array of objects, containing data for navigation links of in-place sorting algorithm.
+ */
 export const inPlaceNavLinks = [
     {
         name: 'Bubble Sort',
-        url: ''
+        url: '#/bubble-sort'
     },
 
     {
         name: 'Selection Sort',
-        url: ''
+        url: '#/selection-sort'
     },
 
     {
         name: 'Insertion Sort',
-        url: ''
+        url: '#/insertion-sort'
     },
 
     {