+++ src/index.js
@@ -1,5 +1,9 @@
-import './styles/index.css';
-import App from './App';
+import './styles/index.css'; 
+import App from './App'; 
 
+/**
+ * The entry point of the application
+ * creates an instance of the App class and runs it
+ */
 const app = new App()
 app.run();