+++ src/views/shared_components/header/Header.js
@@ -5,33 +5,44 @@ import ToggleButton from '../toogle_button/ToggleButton';
 import './header.css';
 
 
+/**
+ * Class representing the Header component
+ */
 class Header {
     
+    /**
+     * Initialize the header element, container element and the root element.
+     */
     constructor(){
+        // Reference to the root element
         this.rootElement = document.firstElementChild;
         
-        // Create the Header Element with container div child.
+        // Create the Header Element with a container div child.
         this.header = document.createElement('header');
         const container = document.createElement('div');
         container.classList.add('container', 'header__container');
         this.header.appendChild(container);
     };
 
+    
      /**
      * Creates the header element and returns it.
      * 
-     * @returns HTMLElement Header element
+     * @returns {HTMLElement} Header Element
      */
      create(){
         // Create The Title
         this.#createHeaderTitle();
         
-        // Create Settings Container With The Settings
+        // Create settings container containing the settings
         this.#createSettingsContainer();
 
         return this.header;
     };
 
+    /**
+     * Creates the title element and appends it to the container element
+     */
     #createHeaderTitle() {
         const title = document.createElement('h1');
         title.classList.add('header__title');
@@ -39,6 +50,10 @@ class Header {
         this.header.firstElementChild.appendChild(title);
     };
 
+    /**
+     * Creates the settings container and adds it to the header element
+     * and creates the burger menu and theme button inside it
+     */
     #createSettingsContainer() {
         const settings = document.createElement('div');
         settings.classList.add('header__settings');
@@ -47,10 +62,16 @@ class Header {
         // Create Burger Menu
         this.#createBurgerMenu(settings);
 
-        // Add a ToggleButton For Changing the Theme
+        // Create Toggle button for toggling between dark and light theme
         this.#createThemeButton(settings);
     };
 
+
+    /**
+    * Creates the theme toggle button and adds it to the settings container.
+    *
+    * @param {HTMLElement} settings - The element that the theme toggle button will be added to.
+    */
     #createThemeButton(settings) {
         const firstIcon = { src: sunIcon, alt: 'Light Theme' };
         const secondIcon = { src: moonIcon, alt: 'Dark Theme' };
@@ -58,20 +79,30 @@ class Header {
             'header__btn',
             firstIcon,
             secondIcon,
-            () => { this.rootElement.classList.toggle('dark-theme'); }).create();
+            () => { this.#handleThemeButtonClick() }).create();
 
         settings.appendChild(themeButton);
     };
 
+     /**
+     * Handles the click event on the theme toggle button.
+     * Toggles the dark-theme class on the root element to switch between themes.
+     */
+    #handleThemeButtonClick() {
+        this.rootElement.classList.toggle('dark-theme');
+    }
+
+    /**
+     * Creates the burger menu button and adds it to the settings container.
+     *
+     * @param {HTMLElement} settings - The element that the burger menu button will be added to.
+     */
     #createBurgerMenu(settings){
         // Create Button
         const burgerBtn = document.createElement('button');
         burgerBtn.classList.add('header__btn', 'header__burger-menu');
         // Add on click listener to open the navbar
-        burgerBtn.onclick = () => {
-            const navbar = document.querySelector('.navbar');
-            navbar.classList.add('active');
-        }
+        burgerBtn.onclick = () => {this.#handleBurgerMenuClick();}
 
         // Add three divs to construct the bars
         for (let i=0; i<=2; i++){
@@ -82,6 +113,15 @@ class Header {
 
         settings.appendChild(burgerBtn);
     };
+
+    /**
+     * Handles the click event on the burger menu button.
+     * Toggles the active class on the navbar to open it.
+     */
+    #handleBurgerMenuClick() {
+        const navbar = document.querySelector('.navbar');
+        navbar.classList.add('active');
+    }
   
 };